Move default event fields implementation to interface module
