+++ lib/basicEvent/index.spec.js
@@ -11,14 +11,14 @@ const mockFetch = {
   fetchAllEvents: () => eventsFixtures.events,
 };
 
-const EventsFields = injectr('../../lib/event/eventFields.js', {
+const BasicEventsFields = injectr('../../lib/basicEvent/index.js', {
   '../fetch': mockFetch,
 });
 
 const EventsSchema = new GraphQLSchema({
   query: new GraphQLObjectType({
     name: 'Query',
-    fields: EventsFields.EventSchemaFields,
+    fields: BasicEventsFields.BasicEventSchemaFields,
   }),
 });
