+++ lib/community/communityFields.js
@@ -10,11 +10,11 @@ import {
   fetchEvent,
 } from '../fetch';
 
-import { EventType } from '../event/eventFields';
+import { BasicEventType } from '../basicEvent';
 import {
   DEFAULT_DISPLAY_LEVEL,
   CommunityEventType,
-} from '../communityEvent/communityEventFields';
+} from '../communityEvent';
 
 function fetchCommunityEventData(eventData) {
   const id = pathOr('', ['data', 'value', 'document', 'id'], eventData);
@@ -42,7 +42,7 @@ const communityType = new GraphQLObjectType({
       description: 'The title of the community mailing list',
     },
     featuredEvent: { // TODO: Test my resolver!
-      type: EventType,
+      type: BasicEventType,
       description: 'The community\'s featured event',
       resolve(community) {
         if (community) {
@@ -61,7 +61,7 @@ const communityType = new GraphQLObjectType({
       type: new GraphQLList(CommunityEventType),
       description: 'The community\'s events',
       resolve(community) {
-        if (community) {
+        if (community && community.events) {
           return community.events.map(fetchCommunityEventData);
         }
         return [];
