--- lib/communityEvent/communityEventFields.spec.js
@@ -1,1 +1,0 @@
-import './communityEventFields'; // It doesn't crash.
