+++ lib/communityEvent/index.js
@@ -2,8 +2,7 @@ import {
   GraphQLObjectType,
   GraphQLEnumType,
 } from 'graphql';
-import { eventTypeFields } from '../event/eventFields';
-import EventInterface from '../interfaces/event';
+import EventInterface, { eventTypeFields } from '../interfaces/event';
 
 export const DEFAULT_DISPLAY_LEVEL = 'Regular';
