+++ lib/communityEvent/index.spec.js
@@ -1,0 +1,1 @@
+import '.'; // It doesn't crash.
