+++ lib/schema.js
@@ -3,7 +3,7 @@ import {
   GraphQLSchema,
 } from 'graphql';
 
-import { EventSchemaFields } from './event/eventFields';
+import { BasicEventSchemaFields } from './basicEvent';
 import { NewsSchemaFields } from './news/newsFields';
 import { CommunitySchemaFields } from './community/communityFields';
 import { SpeakerSchemaFields } from './speaker/speakerFields';
@@ -13,7 +13,7 @@ export const Schema = new GraphQLSchema({
   query: new GraphQLObjectType({
     name: 'Query',
     fields: Object.assign(
-      EventSchemaFields,
+      BasicEventSchemaFields,
       NewsSchemaFields,
       CommunitySchemaFields,
       SpeakerSchemaFields,
